# Various optimizations and API improvements