--- a/src/frontend/mod.rs
+++ b/src/frontend/mod.rs
@@ -1,5 +1,6 @@
 pub mod ast;
 pub mod diag;
+pub mod lexer;
 pub mod parser;
 pub mod printer;
 pub mod sema;