### Optimizations
- On 64 bit platforms CST node indices for the last child node of a rule and the token spans are now stored as 48 bit integers. This allows the `Node` type to be 8 bytes in size as long as the `Token` and `Rule` enums are 1 byte each.
- Tokens are now directly stored in nodes.
- The span calculation for rule nodes no longer recursively visits the child nodes. Instead it just searches from the start and end for the first non-skipped token.
- Space for nodes is now reserved based on the number of tokens when the CST is created.

#### Benchmark
<details>
<summary>Parsing canada.json without printing the CST:</summary>

```
$ hyperfine './lelwel-json canada.json' --warmup 100
``` 
Before:
```
Benchmark 1: ./lelwel-json canada.json
  Time (mean ± σ):       9.4 ms ±   0.4 ms    [User: 5.4 ms, System: 4.1 ms]
  Range (min … max):     8.1 ms …  10.4 ms    210 runs
```
After:
```
Benchmark 1: ./lelwel-json canada.json
  Time (mean ± σ):       7.8 ms ±   0.6 ms    [User: 5.6 ms, System: 2.4 ms]
  Range (min … max):     6.7 ms …  10.3 ms    230 runs
```

</details>
<details>
<summary>Parsing canada.json while printing the CST:</summary>

```
$ hyperfine './lelwel-json canada.json' --warmup 10
``` 
Before:
```
Benchmark 1: ./lelwel-json canada.json
  Time (mean ± σ):     475.5 ms ±  16.4 ms    [User: 340.4 ms, System: 133.9 ms]
  Range (min … max):   453.9 ms … 509.5 ms    10 runs
```
After:
```
Benchmark 1: ./lelwel-json canada.json
  Time (mean ± σ):     410.0 ms ±   8.5 ms    [User: 286.8 ms, System: 122.0 ms]
  Range (min … max):   397.9 ms … 420.9 ms    10 runs
```

</details>

### API Changes
- The `PredicatesAndActions` trait has been renamed to `ParserCallbacks`.
- `Parser::parse` now only takes two arguments `source: &'a str` and `diags: &mut Vec<Diagnostic>`. Tokens and Spans are now calculated in `ParserCallbacks::create_tokens`.
- The `Span` and `CstIndex` types are now defined in `generated.rs`, so they should be deleted from `parser.rs`.
- The `err!` macros is now defined in `generated.rs`, so it should be deleted from `parser.rs`. Diagnostic creation is now defined in `ParserCallacks::create_diagnostic`.
- The `PredicatesAndActions::build` method has been renamed to `ParserCallbacks::create_node`.
- `Cst::get_token` and `Cst::get_rule` have been renamed to `Cst::match_token` and `Cst::match_rule`. Also `Cst::match_rule` now returns a `bool`.
- There is now a `Parser::parse_with_context` function, where the parser context can be specified. For this the `Context` type in `parser.rs` should now be defined as `pub`.
- There is now a `Parser::peek_left` method for peeking into the left context of a parse.

### Other Changes
- The generator now writes separate `lexer.rs` and `parser.rs` files for the default lexer and parser implementation.
- The CST rule nodes are now printed in snake case, so they match the definition in the grammar.
- Renamed any occurrence of `range`/`ranges` to `span`/`spans`.
- The `tokens` vector is now stored in the `Parser` type instead of the `Cst` type.
- Rule methods in `Parser` now use a `rule_` prefix to avoid collisions.
- The generated `tokenize` function now takes a `&str` parameter instead of a `logos::Lexer<Token>`.
- Documentation comments were added to the `pub` methods of the generated types.